# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a RBAC server.",
        "title": "RBAC GO API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/rbac/permission": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC permission"
                ],
                "summary": "Permission 생성",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dblayer.PermissionData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Permission"
                        }
                    }
                }
            }
        },
        "/rbac/permission-assignment": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC permissionAssignment"
                ],
                "summary": "PermissionAssignment 생성",
                "parameters": [
                    {
                        "description": "Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dblayer.PermissionAssignmentData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PermissionAssignment"
                        }
                    }
                }
            }
        },
        "/rbac/permission-assignment/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC permissionAssignment"
                ],
                "summary": "PermissionAssignment 삭제",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PermissionAssignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제된 PermissionAssignment 데이터",
                        "schema": {
                            "$ref": "#/definitions/models.PermissionAssignment"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC permissionAssignment"
                ],
                "summary": "PermissionAssignment Update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PermissionAssignment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update에 사용할 Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dblayer.PermissionAssignmentData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "수정된 PermissionAssignment 데이터",
                        "schema": {
                            "$ref": "#/definitions/models.PermissionAssignment"
                        }
                    }
                }
            }
        },
        "/rbac/permission/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC permission"
                ],
                "summary": "Permission 목록 조회",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dblayer.PermissionsPage"
                        }
                    }
                }
            }
        },
        "/rbac/permission/objects": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC permission"
                ],
                "summary": "Permission 에 해당하는 objects 조회",
                "parameters": [
                    {
                        "description": "Object를 구하는데 필요한 permission 정보",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.PermissionKey"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "허용된 object list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rbac/permission/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC permission"
                ],
                "summary": "Permission 삭제",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Permission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "삭제된 Permission 데이터",
                        "schema": {
                            "$ref": "#/definitions/models.Permission"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "RBAC permission"
                ],
                "summary": "Permission Update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Permission ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update에 사용할 Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dblayer.PermissionData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "수정된 Permission 데이터",
                        "schema": {
                            "$ref": "#/definitions/models.Permission"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dblayer.PermissionAssignmentData": {
            "type": "object",
            "properties": {
                "permission_id": {
                    "type": "integer"
                },
                "role_id": {
                    "type": "integer"
                }
            }
        },
        "dblayer.PermissionData": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                }
            }
        },
...
        "dblayer.PermissionsPage": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "next": {
                    "type": "string"
                },
                "previous": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Permission"
                    }
                }
            }
        },
...
        "models.Permission": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                }
            }
        },
        "models.PermissionAssignment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "permission_id": {
                    "type": "integer"
                },
                "role_id": {
                    "type": "integer"
                }
            }
        },
...
        "rest.PermissionKey": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}